feat: 컬럼 추가로 인한 Add 수정

+++ feature/ingredient-add/src/main/java/com/project/ingredient/barcode/ui/save/SaveIngredientViewModel.kt
@@ -151,7 +151,7 @@ class SaveIngredientViewModel @Inject constructor(
                 val changeList = state.ingredientList.toMutableList()
                 changeList[updateIngredientIndex] = Ingredient(
                     name = igd.name,
-                    count = igd.count.toInt(),
+                    count = igd.count.toDouble(),
                     expirationDate = LocalDate.parse(igd.expirationDate),
                     store = getIndexToIngredientStore(igd.storeSelected),
                     category = getIndexToIngredientCategory(igd.categorySelected),
